RevB1 commit before cleanup
I wanted to be sure I didn't delete something I needed later, so this commit is all files in whatever state. In the next commit I will remove everything I think I no longer need so the repo is a bit cleaner and easier to see what files are important.