Fix SD card footprint and add pull-up resistors
Rev B was using an old and incorrect footprint causing a 3.3v short to ground when an SD card was inserted. Update to use the fixed footprint.

Added pull-up resistors for EN and IO0 (Boot mode). On test board, IO0 appeared to be pulled permanently low. It could be programmed but would not boot. It may only be necessary for EN to be pulled high but added the pull-up for IO0 just incase.